This merges all the changes from branch 'unittests' and deletes oops/.vscode
