+++ hosts/unittester.py
@@ -4,8 +4,9 @@
 
 import unittest
 
-from cassini.iss import Test_Cassini_ISS
-from hst         import Test_HST
+from juno.junocam   import Test_Juno_Junocam
+#from cassini.iss    import Test_Cassini_ISS
+#from hst            import Test_HST
 
 ########################################
 if __name__ == '__main__':
