+++ oops/fov/__init__.py
@@ -330,7 +330,8 @@ class FOV(object):
         """
 
         xy_pair = self.xy_from_los(los, derivs=derivs)
-        return self.uv_from_xyt(xy_pair, tfrac, time, derivs=derivs, **keywords)
+        return self.uv_from_xyt(xy_pair,
+                            tfrac=tfrac, time=time, derivs=derivs, **keywords)
 
     #===========================================================================
     def uv_from_los(self, los, derivs=False, **keywords):
@@ -471,7 +472,7 @@ class FOV(object):
                         full FOV.
         """
 
-        uv = self.uv_from_xyt(xy_pair, tfrac, derivs=False, **keywords)
+        uv = self.uv_from_xyt(xy_pair, tfrac=tfrac, time=time, derivs=False, **keywords)
         return self.uv_is_outside(uv, inclusive, uv_min, uv_max)
 
     #===========================================================================
