+++ oops/path/spicepath.py
@@ -266,7 +266,7 @@ class Test_SpicePath(unittest.TestCase):
         Path.reset_registry()
         Frame.reset_registry()
 
-        _     = SpicePath("SUN", "SSB")
+        _sun  = SpicePath("SUN", "SSB")
         earth = SpicePath("EARTH", "SSB")
         moon  = SpicePath("MOON", "EARTH")
